2022/requirements.txt
pycairo
svg.path

pylint
mypy